1500-Moved unit tests to tests/unit and integration tests to tests/integration. Change circle ci config to run all onestop-python-client tests.
