+++ .circleci/config.yml
@@ -107,10 +107,16 @@ jobs:
           app-dir: ./onestop-python-client
           pkg-manager: pip
       - run:
-          name: "Run util tests"
+          name: "Run unit tests"
           command: >
             cd onestop-python-client/;
-            python -m unittest tests/util/*.py
+            python -m unittest discover -s test/unit
+
+      - run:
+          name: "Run integration tests"
+          command: >
+            cd onestop-python-client/;
+            python -m unittest discover -s test/integration
 
 orbs:
   slack: circleci/slack@3.4.2
