+++ onestop-python-client/test/unit/util/test_S3MessageAdapter.py
