+++ onestop-python-client/test/unit/util/test_SqsConsumer.py
