--- onestop-python-client/tests/util/IntegrationTest.py
@@ -1,1 +1,0 @@
-#TBD
